Rename ./config.js to ./particle.root.config.js
