+++ apps/drupal/particle.app.config.js
@@ -8,7 +8,7 @@ const {
   PATH_DIST,
   ASSETS_BUNDLE_FOLDER,
   ASSETS_ATOMIC_FOLDER,
-} = require('../../config');
+} = require('../../particle.root.config');
 
 // Used as folder name within PATH_DIST, does not have to be folder name of app
 const APP_NAME = 'app-drupal';
