+++ particle.root.config.js
